requirements.txt
fastapi
uvicorn
streamlit
pymilvus==2.4.4
langchain
langchain-community
ollama
requests
python-dotenv
tiktoken 